Go back to `.. == EOF || ... == EOF` pattern
For code which propagates the write error as an error,
this is the more universal pattern. The other pattern,
`... != EOF && ... != EOF`, is a little more readable,
but doesn't generalize as well, yet remains better in
code which swallows the write error, such as fatal
error printing code, because a command-line tool has
nothing good to do with that error anyway.